+++ json/currencyUpdater30_04_2022.json
@@ -1,0 +1,1 @@
+[{"currencyName":"AUD","currencyScale":1,"Ratio":1.8863,"currencyId":440},{"currencyName":"AMD","currencyScale":1000,"Ratio":5.752,"currencyId":510},{"currencyName":"BGN","currencyScale":1,"Ratio":1.4267,"currencyId":441},{"currencyName":"UAH","currencyScale":100,"Ratio":8.7716,"currencyId":449},{"currencyName":"DKK","currencyScale":10,"Ratio":3.75,"currencyId":450},{"currencyName":"USD","currencyScale":1,"Ratio":2.6534,"currencyId":431},{"currencyName":"EUR","currencyScale":1,"Ratio":2.7896,"currencyId":451},{"currencyName":"PLN","currencyScale":10,"Ratio":5.9465,"currencyId":452},{"currencyName":"JPY","currencyScale":100,"Ratio":2.0349,"currencyId":508},{"currencyName":"IRR","currencyScale":100000,"Ratio":6.3176,"currencyId":461},{"currencyName":"ISK","currencyScale":100,"Ratio":2.0222,"currencyId":453},{"currencyName":"CAD","currencyScale":1,"Ratio":2.0662,"currencyId":371},{"currencyName":"CNY","currencyScale":10,"Ratio":4.0141,"currencyId":462},{"currencyName":"KWD","currencyScale":1,"Ratio":8.6534,"currencyId":394},{"currencyName":"MDL","currencyScale":10,"Ratio":1.4289,"currencyId":454},{"currencyName":"NZD","currencyScale":1,"Ratio":1.7226,"currencyId":448},{"currencyName":"NOK","currencyScale":10,"Ratio":2.8289,"currencyId":455},{"currencyName":"RUB","currencyScale":100,"Ratio":3.6694,"currencyId":456},{"currencyName":"XDR","currencyScale":1,"Ratio":3.5786,"currencyId":457},{"currencyName":"SGD","currencyScale":1,"Ratio":1.9153,"currencyId":421},{"currencyName":"KGS","currencyScale":100,"Ratio":3.1529,"currencyId":458},{"currencyName":"KZT","currencyScale":1000,"Ratio":5.958,"currencyId":459},{"currencyName":"TRY","currencyScale":10,"Ratio":1.7923,"currencyId":460},{"currencyName":"GBP","currencyScale":1,"Ratio":3.319,"currencyId":429},{"currencyName":"CZK","currencyScale":100,"Ratio":11.3785,"currencyId":463},{"currencyName":"SEK","currencyScale":10,"Ratio":2.7064,"currencyId":464},{"currencyName":"CHF","currencyScale":1,"Ratio":2.731,"currencyId":426}]
