+++ json/testapi.json
@@ -1,0 +1,1 @@
+[{"Cur_ID":440,"Cur_Name":"Австралийский доллар","Cur_OfficialRate":1.8863,"Cur_Abbreviation":"AUD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":510,"Cur_Name":"Армянских драмов","Cur_OfficialRate":5.752,"Cur_Abbreviation":"AMD","Date":"2022-05-01T00:00:00","Cur_Scale":1000},{"Cur_ID":441,"Cur_Name":"Болгарский лев","Cur_OfficialRate":1.4267,"Cur_Abbreviation":"BGN","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":449,"Cur_Name":"Гривен","Cur_OfficialRate":8.7716,"Cur_Abbreviation":"UAH","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":450,"Cur_Name":"Датских крон","Cur_OfficialRate":3.75,"Cur_Abbreviation":"DKK","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":431,"Cur_Name":"Доллар США","Cur_OfficialRate":2.6534,"Cur_Abbreviation":"USD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":451,"Cur_Name":"Евро","Cur_OfficialRate":2.7896,"Cur_Abbreviation":"EUR","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":452,"Cur_Name":"Злотых","Cur_OfficialRate":5.9465,"Cur_Abbreviation":"PLN","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":508,"Cur_Name":"Иен","Cur_OfficialRate":2.0349,"Cur_Abbreviation":"JPY","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":461,"Cur_Name":"Иранских риалов","Cur_OfficialRate":6.3176,"Cur_Abbreviation":"IRR","Date":"2022-05-01T00:00:00","Cur_Scale":100000},{"Cur_ID":453,"Cur_Name":"Исландских крон","Cur_OfficialRate":2.0222,"Cur_Abbreviation":"ISK","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":371,"Cur_Name":"Канадский доллар","Cur_OfficialRate":2.0662,"Cur_Abbreviation":"CAD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":462,"Cur_Name":"Китайских юаней","Cur_OfficialRate":4.0141,"Cur_Abbreviation":"CNY","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":394,"Cur_Name":"Кувейтский динар","Cur_OfficialRate":8.6534,"Cur_Abbreviation":"KWD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":454,"Cur_Name":"Молдавских леев","Cur_OfficialRate":1.4289,"Cur_Abbreviation":"MDL","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":448,"Cur_Name":"Новозеландский доллар","Cur_OfficialRate":1.7226,"Cur_Abbreviation":"NZD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":455,"Cur_Name":"Норвежских крон","Cur_OfficialRate":2.8289,"Cur_Abbreviation":"NOK","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":456,"Cur_Name":"Российских рублей","Cur_OfficialRate":3.6694,"Cur_Abbreviation":"RUB","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":457,"Cur_Name":"СДР (Специальные права заимствования)","Cur_OfficialRate":3.5786,"Cur_Abbreviation":"XDR","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":421,"Cur_Name":"Сингапурcкий доллар","Cur_OfficialRate":1.9153,"Cur_Abbreviation":"SGD","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":458,"Cur_Name":"Сомов","Cur_OfficialRate":3.1529,"Cur_Abbreviation":"KGS","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":459,"Cur_Name":"Тенге","Cur_OfficialRate":5.958,"Cur_Abbreviation":"KZT","Date":"2022-05-01T00:00:00","Cur_Scale":1000},{"Cur_ID":460,"Cur_Name":"Турецких лир","Cur_OfficialRate":1.7923,"Cur_Abbreviation":"TRY","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":429,"Cur_Name":"Фунт стерлингов","Cur_OfficialRate":3.319,"Cur_Abbreviation":"GBP","Date":"2022-05-01T00:00:00","Cur_Scale":1},{"Cur_ID":463,"Cur_Name":"Чешских крон","Cur_OfficialRate":11.3785,"Cur_Abbreviation":"CZK","Date":"2022-05-01T00:00:00","Cur_Scale":100},{"Cur_ID":464,"Cur_Name":"Шведских крон","Cur_OfficialRate":2.7064,"Cur_Abbreviation":"SEK","Date":"2022-05-01T00:00:00","Cur_Scale":10},{"Cur_ID":426,"Cur_Name":"Швейцарский франк","Cur_OfficialRate":2.731,"Cur_Abbreviation":"CHF","Date":"2022-05-01T00:00:00","Cur_Scale":1}]
