--- src/wallet/money/currencyUnit/builders/CurrencyUnitStorageBuilder.java
@@ -1,8 +1,0 @@
-package wallet.money.currencyUnit.builders;
-
-public interface CurrencyUnitStorageBuilder {
-
-    void reset();
-    void buildCurrencyUnit(String currencyString);
-
-}
