+++ src/wallet/money/currencyUnit/interfaces/CurrencyUnitStorage.java
@@ -1,0 +1,11 @@
+package wallet.money.currencyUnit.interfaces;
+
+import wallet.money.currencyUnit.StrictCurrencyUnit;
+
+public interface CurrencyUnitStorage {
+
+    boolean isCurrencyExists(String currencyString);
+    StrictCurrencyUnit getCurrencyUnitByCurrencyString(String currencyString);
+    StrictCurrencyUnit getCurrencyUnitByCurrencyID(long currencyId);
+
+}
