+++ src/wallet/money/currencyUnit/interfaces/CurrencyUnitStorageBuilder.java
@@ -1,0 +1,8 @@
+package wallet.money.currencyUnit.interfaces;
+
+public interface CurrencyUnitStorageBuilder {
+
+    void reset();
+    void buildCurrencyUnit(String currencyString);
+
+}
