+++ src/wallet/money/currencyUnit/interfaces/CurrencyUpdaterBuilder.java
@@ -1,0 +1,8 @@
+package wallet.money.currencyUnit.interfaces;
+
+public interface CurrencyUpdaterBuilder {
+
+    void reset();
+    void buildCurrency(String currencyString);
+
+}
