--- src/wallet/money/currencyUpdaters/builders/CurrencyUpdaterBuilder.java
@@ -1,8 +1,0 @@
-package wallet.money.currencyUpdaters.builders;
-
-public interface CurrencyUpdaterBuilder {
-
-    void reset();
-    void buildCurrency(String currencyString);
-
-}
